requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0

# SQLAlchemy + drivere
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
psycopg[binary]==3.1.19

# Migrații
alembic==1.13.1

# Config & env
python-dotenv==1.0.1
pydantic-settings==2.2.1

# Pydantic (inclusiv validarea email-urilor)
pydantic[email]==2.7.1

# Autentificare și securitate
passlib[bcrypt]==1.7.4
pyjwt==2.8.0
